rockchip: bob: Build depthcharge partition image
This is using the KERNEL_START and KERNEL_SIZE values from depthcharge's
firmware-gru-8785.B branch. Also updates SYS_TEXT_BASE to KERNEL_START +
0x100, like others.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Series-name: binman-depthcharge
Commit-notes:
This is the same as what I'm doing on a chromebook_kevin, but U-Boot
doesn't really work in any form on it yet.
END