doc: Document binman-built images in README.chromium-chainload
Some of this is copied from the text below it with the hope of
eventually replacing it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Series-name: binman-depthcharge
Commit-notes:
I think the older parts (that I've automated now) can be removed, but
wanted to keep this diff very clean and wait until someone can really
test the images.
END