rockchip: bob: Enable configs for the display
The board uses an eDP panel compatible with simple-panel and rk3399-edp.
Its backlight is controlled by the Chrome OS Embedded Controller PWM, so
enable that as well.

The rockchip video driver needs to assert/deassert some resets, so also
enable the reset controller. RESET_ROCKCHIP defaults to y for this board
when DM_RESET=y, so it's enough to set that.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>